Refactoring changes. Paving the way for the password feature.
