+++ APIQuiz/Models/Player.cs
@@ -8,7 +8,7 @@ namespace APIQuiz.Models
         public int Id { get; set; }
         public int Score { get; set; }
         public string Name { get; set; }
-        public void CopyUpdatedDataFrom(Player updatedPlayer)
+        public void CopyDataFrom(Player updatedPlayer)
         {
             Name = updatedPlayer.Name;
         }
