+++ APIQuizTests/Services/GameServiceTests.cs
@@ -1,5 +1,4 @@
-﻿using APIQuiz.Services;
-using Xunit;
+﻿using Xunit;
 using System.Threading.Tasks;
 using APIQuiz.Models;
